Insert a notification mechanism so npiet can notify observers when things change internally.
This is difficult because we are in C. As such we don't have anything like signals/slots. The closest we have is function pointers and callbacks. So I've implemented a hopefully minimally intrusive notification callback system into npiet so an observer can be notified of events during the process of execution.